Better-er support for firefox
Firefox addons need to have a unique add-on ID. The add-on ID is specified via `browser_specific_settings.gecko.id` in the manifest. #616 consistently set the add-on ID to "example@example.com", which won't work for anyone else who wants to publish an add-on.

Make the add-on ID configurable via the core manifest.js, since chrome extensions don't seem to care about the extra manifest key.